unify timeout, drop and normal response handling
Makes timeouts, drops and normal responses follow the same code
path, and ensures that we won't get IllegalStateExceptions from
AsyncContext in 'normal' operation.
